pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "fastapi>=0.115.12",
    "pydantic>=2.11.4",
    "pymongo==4.10.1",
    "python-dotenv>=1.1.0",
]
